+++ colour_text/main.cpp
@@ -8,7 +8,7 @@ using namespace std;
 int main() {
 
 	// Configure the colours and titles.
-	const size_t SIZE = 8;
+	const unsigned SIZE = 8;
 	char colours[] = {'X', 'R', 'G', 'Y', 'B', 'M', 'C', 'W'};
 	string title[] = {
 		"grey",
@@ -26,9 +26,9 @@ int main() {
 	string gamma;
 	char colour;
 	cout << "Colour texts:" << endl;
-	for (uint i = 0; i < 2; i++) {
+	for (unsigned i = 0; i < 2; i++) {
 		gamma = (i == 0 ? "bright " : "dark ");
-		for (uint j = 0; j < SIZE; j++) {
+		for (unsigned j = 0; j < SIZE; j++) {
 			colour = (i == 0 ? tolower(colours[j]) : colours[j]);
 			ss << colour << " --> " << gamma << title[j];
 			cout << colourText(ss.str(), colour) << endl;
@@ -38,9 +38,9 @@ int main() {
 
 	// Starting and ending specific colours.
 	cout << "\nColour start and end:" << endl;
-	for (uint i = 0; i < 2; i++) {
+	for (unsigned i = 0; i < 2; i++) {
 		gamma = (i == 0 ? "bright " : "dark ");
-		for (uint j = 0; j < SIZE; j++) {
+		for (unsigned j = 0; j < SIZE; j++) {
 			colour = (i == 0 ? tolower(colours[j]) : colours[j]);
 			startColour(colour);
 			cout << colour;
